backend/requirements.txt
flask
python-nmap
flask-cors
requests